Extract vertices issue dev (#2314)
* pgr_extractVertices
  * [pgtap] adding tests of failure
  * [sql] fix
  * [doc] updating news
* fix sponsors link

+++ sql/topology/extractVertices.sql
@@ -228,12 +228,12 @@ BEGIN
           FULL OUTER JOIN agg_in USING (vid)
         )
 
-        SELECT vid AS id, in_edges, out_edges, NULL::FLOAT, NULL::FLOAT, NULL::geometry
+        SELECT vid::BIGINT AS id, in_edges, out_edges, NULL::FLOAT, NULL::FLOAT, NULL::geometry
         FROM the_points$q$;
 
 
     ELSIF has_source AND NOT has_id THEN
-      -- SELECT id, source, target
+      -- SELECT source, target
       query := $q$
         WITH
 
@@ -248,7 +248,7 @@ BEGIN
           SELECT target FROM main_sql
         )
 
-        SELECT DISTINCT vid AS id, NULL::BIGINT[], NULL::BIGINT[], NULL::FLOAT, NULL::FLOAT, NULL::geometry
+        SELECT DISTINCT vid::BIGINT AS id, NULL::BIGINT[], NULL::BIGINT[], NULL::FLOAT, NULL::FLOAT, NULL::geometry
         FROM the_points$q$;
 
     ELSE
